Debugging / Removing Excess Code
Removing unnecessary code / fixing the problems that arose from forgetting to have a min-height on the  <main> element (a mistake that resulted in many hours searching for a quick fix on a wild-goose chase... min-height was all that was missing). This sessions was to correct the mess that resulted from the previous mistake. (The specific issue had been an uncooperative footer section that wouldn't remain within its own bounds.)