First Media Query Refactoring Complete
Refactored spacing between elements for media query less than 451px.  Removed some commented out code that should no longer be needed. Added a title comment to a block of code on HTML. Only difference between original and clone is a min-width that keeps the page from collapsing below a certain amount (opinion: this design is an improvement from the original...)