Fixed margin on "Get App" buttons
There was margin on the images used for the get-app buttons, which was causing extra white space around the anchor tag of the button in which the pointer was changing to the pointing hand. Removed the margin from the images, and placed it back on the anchor tag to eliminate that extra "pointer" space.